+++ examples/simple.rs
@@ -5,7 +5,7 @@
 
 use actix_web::web::ReqData;
 use actix_web::{middleware, web, App, HttpResponse, HttpServer, Responder};
-use actix_web_middleware_keycloak_auth::{Claims, DecodingKey, KeycloakAuth};
+use actix_web_middleware_keycloak_auth::{Claims, DecodingKey, KeycloakAuth, Role};
 
 const KEYCLOAK_PK: &str = "-----BEGIN PUBLIC KEY-----
 MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAnzyis1ZjfNB0bBgKFMSv
@@ -57,7 +57,9 @@ async fn main() -> std::io::Result<()> {
         let keycloak_auth = KeycloakAuth {
             detailed_responses: true,
             keycloak_oid_public_key: DecodingKey::from_rsa_pem(KEYCLOAK_PK.as_bytes()).unwrap(),
-            required_roles: vec!["test".to_owned()],
+            required_roles: vec![Role::Realm {
+                role: "test".to_owned(),
+            }],
         };
 
         App::new()