+++ src/roles.rs
@@ -7,15 +7,22 @@ use jsonwebtoken::TokenData;
 use log::debug;
 
 use crate::errors::AuthError;
-use crate::Claims;
+use crate::{Claims, Role};
 
 pub fn check_roles(
     token: TokenData<Claims>,
-    required_roles: &[String],
+    required_roles: &[Role],
 ) -> Result<TokenData<Claims>, AuthError> {
     let roles = token.claims.roles();
 
-    debug!("JWT contains roles: {}", &roles.join(", "));
+    debug!(
+        "JWT contains roles: {}",
+        &roles
+            .iter()
+            .map(|r| r.to_string())
+            .collect::<Vec<String>>()
+            .join(", ")
+    );
 
     let mut missing_roles = vec![];
     for role in required_roles {
@@ -35,9 +42,11 @@ pub fn check_roles(
 mod tests {
     use chrono::Utc;
     use jsonwebtoken::{Algorithm, Header};
+    use std::collections::HashMap;
+    use std::iter::FromIterator;
 
     use super::*;
-    use crate::RealmAccess;
+    use crate::Access;
 
     #[test]
     fn no_required_no_provided() {
@@ -47,6 +56,7 @@ mod tests {
                 sub: "".to_owned(),
                 exp: Utc::now(),
                 realm_access: None,
+                resource_access: None,
             },
         };
         let required_roles = &[];
@@ -61,9 +71,10 @@ mod tests {
             claims: Claims {
                 sub: "".to_owned(),
                 exp: Utc::now(),
-                realm_access: Some(RealmAccess {
+                realm_access: Some(Access {
                     roles: vec!["test1".to_owned(), "test2".to_owned()],
                 }),
+                resource_access: None,
             },
         };
         let required_roles = &[];
@@ -79,15 +90,30 @@ mod tests {
                 sub: "".to_owned(),
                 exp: Utc::now(),
                 realm_access: None,
+                resource_access: None,
             },
         };
-        let required_roles = &["test1".to_owned(), "test2".to_owned()];
+        let required_roles = &[
+            Role::Realm {
+                role: "test1".to_owned(),
+            },
+            Role::Realm {
+                role: "test2".to_owned(),
+            },
+        ];
 
         let result = check_roles(token, required_roles);
         assert!(result.is_err());
         assert_eq!(
             result.unwrap_err(),
-            AuthError::MissingRoles(vec!["test1".to_owned(), "test2".to_owned()])
+            AuthError::MissingRoles(vec![
+                Role::Realm {
+                    role: "test1".to_owned()
+                },
+                Role::Realm {
+                    role: "test2".to_owned()
+                }
+            ])
         );
     }
 
@@ -98,18 +124,28 @@ mod tests {
             claims: Claims {
                 sub: "".to_owned(),
                 exp: Utc::now(),
-                realm_access: Some(RealmAccess {
+                realm_access: Some(Access {
                     roles: vec!["test2".to_owned()],
                 }),
+                resource_access: None,
             },
         };
-        let required_roles = &["test1".to_owned(), "test2".to_owned()];
+        let required_roles = &[
+            Role::Realm {
+                role: "test1".to_owned(),
+            },
+            Role::Realm {
+                role: "test2".to_owned(),
+            },
+        ];
 
         let result = check_roles(token, required_roles);
         assert!(result.is_err());
         assert_eq!(
             result.unwrap_err(),
-            AuthError::MissingRoles(vec!["test1".to_owned(),])
+            AuthError::MissingRoles(vec![Role::Realm {
+                role: "test1".to_owned()
+            }])
         );
     }
 
@@ -120,12 +156,20 @@ mod tests {
             claims: Claims {
                 sub: "".to_owned(),
                 exp: Utc::now(),
-                realm_access: Some(RealmAccess {
+                realm_access: Some(Access {
                     roles: vec!["test1".to_owned(), "test2".to_owned()],
                 }),
+                resource_access: None,
             },
         };
-        let required_roles = &["test1".to_owned(), "test2".to_owned()];
+        let required_roles = &[
+            Role::Realm {
+                role: "test1".to_owned(),
+            },
+            Role::Realm {
+                role: "test2".to_owned(),
+            },
+        ];
 
         assert!(check_roles(token, required_roles).is_ok());
     }
@@ -137,12 +181,114 @@ mod tests {
             claims: Claims {
                 sub: "".to_owned(),
                 exp: Utc::now(),
-                realm_access: Some(RealmAccess {
+                realm_access: Some(Access {
                     roles: vec!["test1".to_owned(), "test2".to_owned(), "test3".to_owned()],
                 }),
+                resource_access: None,
             },
         };
-        let required_roles = &["test1".to_owned(), "test2".to_owned()];
+        let required_roles = &[
+            Role::Realm {
+                role: "test1".to_owned(),
+            },
+            Role::Realm {
+                role: "test2".to_owned(),
+            },
+        ];
+
+        assert!(check_roles(token, required_roles).is_ok());
+    }
+
+    #[test]
+    fn client_roles() {
+        let token = TokenData {
+            header: Header::new(Algorithm::RS256),
+            claims: Claims {
+                sub: "".to_owned(),
+                exp: Utc::now(),
+                realm_access: Some(Access {
+                    roles: vec!["test1".to_owned(), "test2".to_owned()],
+                }),
+                resource_access: Some(HashMap::from_iter(vec![
+                    (
+                        "client1".to_owned(),
+                        Access {
+                            roles: vec!["role1".to_owned(), "role2".to_owned()],
+                        },
+                    ),
+                    (
+                        "client2".to_owned(),
+                        Access {
+                            roles: vec!["role3".to_owned()],
+                        },
+                    ),
+                    ("client3".to_owned(), Access { roles: vec![] }),
+                ])),
+            },
+        };
+        let required_roles = &[
+            Role::Realm {
+                role: "test1".to_owned(),
+            },
+            Role::Realm {
+                role: "test2".to_owned(),
+            },
+            Role::Client {
+                client: "client1".to_owned(),
+                role: "role1".to_owned(),
+            },
+            Role::Client {
+                client: "client1".to_owned(),
+                role: "role2".to_owned(),
+            },
+            Role::Client {
+                client: "client2".to_owned(),
+                role: "role3".to_owned(),
+            },
+        ];
+
+        assert!(check_roles(token, required_roles).is_ok());
+    }
+
+    #[test]
+    fn both_realm_and_client_roles() {
+        let token = TokenData {
+            header: Header::new(Algorithm::RS256),
+            claims: Claims {
+                sub: "".to_owned(),
+                exp: Utc::now(),
+                realm_access: None,
+                resource_access: Some(HashMap::from_iter(vec![
+                    (
+                        "client1".to_owned(),
+                        Access {
+                            roles: vec!["role1".to_owned(), "role2".to_owned()],
+                        },
+                    ),
+                    (
+                        "client2".to_owned(),
+                        Access {
+                            roles: vec!["role3".to_owned()],
+                        },
+                    ),
+                    ("client3".to_owned(), Access { roles: vec![] }),
+                ])),
+            },
+        };
+        let required_roles = &[
+            Role::Client {
+                client: "client1".to_owned(),
+                role: "role1".to_owned(),
+            },
+            Role::Client {
+                client: "client1".to_owned(),
+                role: "role2".to_owned(),
+            },
+            Role::Client {
+                client: "client2".to_owned(),
+                role: "role3".to_owned(),
+            },
+        ];
 
         assert!(check_roles(token, required_roles).is_ok());
     }