+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+- support client roles
+
 ## 0.1.0
 
 Initial release